change app.jsx path
